build: Replace build system with CMake
This patch replaces the previous build system (based on Ninja files
generated from a Python script) with CMake.

For a quick walk-through, see `HACKING.md`.

I have tested this build system with all combinations of:

- Clang and GCC
- CMake 3.5.2 and CMake 3.11.2

The most important characteristic of this new build system is how
dependencies are handled.

Previously, dependencies of Seastar that were not readily available as
system packages were provided through Git submodules. Part of Seastar's
build involved launching external tools to build these dependencies.

In this build, we specify with CMake the names and versions of all
dependencies, but we do not bundle them (except for DPDK. See below).

In this way, the dependencies can be supplied from arbitrary sources.
These could include system packages, custom locations on disk, or
package managers like Conan.

One option for supplying dependencies included in this patch is the
`cooking.sh` script, part of the `cmake-cooking` project. This script
creates a local development environment containing all (or some)
dependencies by fetching their source code, compiling them, and
installing them into a project-specific location. More instructions for
using `cooking.sh` are available in the new `HACKING.md` file.

Because the build system only checks that dependencies are satisfied,
when Seastar is installed via -- for example, `ninja install` -- the
installed files are relocatable to any system which includes the
dependencies and for which the installation destination exists.

As part of this build system, we also generate a pkg-config
specification which is installed alongside the CMake files necessary for
consumers to use Seastar.

This new build system also includes the new `test_dist` target. This
test verifies that clients can build applications against Seastar with
the files that the build-system installs. The test uses `cooking.sh` to
manage all dependencies and checks both "native" CMake support and also
compilation with pkg-config.

Finally, a brief note on Git submodules and dependencies.

When a dependency is tracked at the level of a Git commit (like DPDK),
then it is managed as a submodule in `extern` (e.g., `extern/dpkd`). In
this case, it can still be included with `cmake-cooking` using the
`SOURCE_DIR` argument. Otherwise, it is best to track dependencies based
on their released versions (in order to maximize compatibility and
integration with other platforms and systems) and not use submodules.